Working on adding the compression factory to the indexer and the search engine.  Note that the carryover-12 fails when the postings list length is the same as the number of documents.
git-svn-id: https://groke.otago.ac.nz/ant/ant@211 7489a7b4-4588-4892-8acd-774bacfab51c

+++ source/compression_factory.h
@@ -18,10 +18,10 @@ private:
 	static ANT_compress *technique[];
 	static char *technique_name[];
 	static long number_of_techniques;
-	long *histogram;
+	long long *histogram;
 	
 public:
-	ANT_compression_factory() { histogram = 1 + (long *)memset(new long[number_of_techniques + 1], 0, sizeof(*histogram) * (number_of_techniques + 1)); }
+	ANT_compression_factory() { histogram = 1 + (long long *)memset(new long long[number_of_techniques + 1], 0, sizeof(*histogram) * (number_of_techniques + 1)); }
 	virtual ~ANT_compression_factory() {}
 
 	virtual long long compress(unsigned char *destination, long long destination_length, ANT_compressable_integer *source, long long source_integers);
